frontend/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pyOpenSSL==23.2.0
urllib3==2.0.4
